Refactor main class argument parser
