+++ src/main/java/org/dita/dost/invoker/Main.java
@@ -27,8 +27,6 @@
 package org.dita.dost.invoker;
 
 import com.google.common.base.Strings;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
 import org.apache.tools.ant.*;
 import org.apache.tools.ant.input.DefaultInputHandler;
 import org.apache.tools.ant.input.InputHandler;
@@ -42,8 +40,6 @@ import org.dita.dost.project.Project.Context;
 import org.dita.dost.project.Project.Publication;
 import org.dita.dost.project.ProjectFactory;
 import org.dita.dost.util.Configuration;
-import org.dita.dost.util.XMLUtils;
-import org.w3c.dom.Element;
 
 import java.io.*;
 import java.net.URI;
@@ -52,10 +48,9 @@ import java.nio.file.Paths;
 import java.util.*;
 import java.util.stream.Collectors;
 
+import static org.dita.dost.invoker.Arguments.*;
 import static org.dita.dost.util.Configuration.transtypes;
 import static org.dita.dost.util.Constants.ANT_TEMP_DIR;
-import static org.dita.dost.util.XMLUtils.getChildElements;
-import static org.dita.dost.util.XMLUtils.toList;
 
 /**
  * Command line entry point into DITA-OT. This class is entered via the canonical
@@ -64,265 +59,10 @@ import static org.dita.dost.util.XMLUtils.toList;
  */
 public class Main extends org.apache.tools.ant.Main implements AntMain {
 
-    private boolean useColor;
-
-    private static abstract class Argument {
-        final String property;
-
-        Argument(final String property) {
-            this.property = property;
-        }
-
-        abstract String getValue(final String value);
-    }
-
-    private static class StringArgument extends Argument {
-        StringArgument(final String property) {
-            super(property);
-        }
-
-        @Override
-        String getValue(final String value) {
-            return value;
-        }
-    }
-
-    private static class BooleanArgument extends Argument {
-        final String trueValue;
-        final String falseValue;
-
-        BooleanArgument(final String property, final String trueValue, final String falseValue) {
-            super(property);
-            this.trueValue = trueValue;
-            this.falseValue = falseValue;
-        }
-
-        @Override
-        String getValue(final String value) {
-            switch (value.toLowerCase()) {
-                case "true":
-                case "yes":
-                case "on":
-                case "1":
-                    return trueValue;
-                default:
-                    return falseValue;
-            }
-        }
-    }
-
-    private static class EnumArgument extends Argument {
-        final Set<String> values;
-
-        EnumArgument(final String property, final Set<String> values) {
-            super(property);
-            this.values = values;
-        }
-
-        @Override
-        String getValue(final String value) {
-            if (!values.contains(value)) {
-                throw new BuildException("Invalid value for property " + property + ": " + value);
-            }
-            return value;
-        }
-    }
-
-    private static class FileArgument extends Argument {
-        FileArgument(final String property) {
-            super(property);
-        }
-
-        @Override
-        String getValue(final String value) {
-            return new File(value).getPath();
-        }
-    }
-
-    private static class AbsoluteFileArgument extends Argument {
-        AbsoluteFileArgument(final String property) {
-            super(property);
-        }
-
-        @Override
-        String getValue(final String value) {
-            return new File(value).getAbsolutePath();
-        }
-    }
-
-    private static class AbsoluteFileListArgument extends Argument {
-        AbsoluteFileListArgument(final String property) {
-            super(property);
-        }
-
-        @Override
-        String getValue(final String value) {
-            return Arrays.stream(value.split(File.pathSeparator))
-                    .map(oneFile -> new File(oneFile).getAbsolutePath())
-                    .collect(Collectors.joining(File.pathSeparator));
-        }
-    }
-
-    private static class FileOrUriArgument extends Argument {
-        FileOrUriArgument(final String property) {
-            super(property);
-        }
-
-        @Override
-        String getValue(final String value) {
-            final File f = new File(value);
-            if (f.exists()) {
-                return f.getAbsolutePath();
-            } else {
-                return value;
-            }
-        }
-    }
-
-    /**
-     * A Set of args are are handled by the launcher and should not be seen by
-     * Main.
-     */
-    private static final Set<String> LAUNCH_COMMANDS = new HashSet<>();
-
-    static {
-        LAUNCH_COMMANDS.add("-lib");
-        LAUNCH_COMMANDS.add("-cp");
-        LAUNCH_COMMANDS.add("-noclasspath");
-        LAUNCH_COMMANDS.add("-nouserlib");
-        LAUNCH_COMMANDS.add("-main");
-    }
-
-    private static final Map<String, Argument> ARGUMENTS = new HashMap<>();
-
-    static {
-        ARGUMENTS.put("-f", new StringArgument("transtype"));
-        ARGUMENTS.put("--format", new StringArgument("transtype"));
-        ARGUMENTS.put("--transtype", new StringArgument("transtype"));
-        ARGUMENTS.put("--deliverable", new StringArgument("project.deliverable"));
-        ARGUMENTS.put("-i", new FileOrUriArgument("args.input"));
-        ARGUMENTS.put("--input", new FileOrUriArgument("args.input"));
-        ARGUMENTS.put("-o", new AbsoluteFileArgument("output.dir"));
-        ARGUMENTS.put("--output", new AbsoluteFileArgument("output.dir"));
-        ARGUMENTS.put("--filter", new AbsoluteFileListArgument("args.filter"));
-        ARGUMENTS.put("-t", new AbsoluteFileArgument(ANT_TEMP_DIR));
-        ARGUMENTS.put("--temp", new AbsoluteFileArgument(ANT_TEMP_DIR));
-        ARGUMENTS.put("-p", new AbsoluteFileArgument("project.file"));
-        ARGUMENTS.put("--project", new AbsoluteFileArgument("project.file"));
-        addSingleHyphenOptions(ARGUMENTS);
-    }
-
-    private static void addSingleHyphenOptions(final Map<String, Argument> args) {
-        for (final Map.Entry<String, Argument> e : new HashMap<>(args).entrySet()) {
-            if (e.getKey().startsWith("--")) {
-                args.put(e.getKey().substring(1), e.getValue());
-            }
-        }
-    }
-
-    private static Map<String, Argument> PLUGIN_ARGUMENTS;
-
-    // Lazy load parameters
-    private synchronized Map<String, Argument> getPluginArguments() {
-        if (PLUGIN_ARGUMENTS == null) {
-            final List<Element> params = toList(Plugins.getPluginConfiguration().getElementsByTagName("param"));
-            PLUGIN_ARGUMENTS = params.stream()
-                    .map(Main::getArgument)
-                    .collect(Collectors.toMap(
-                            arg -> ("--" + arg.property),
-                            arg -> arg,
-                            Main::mergeArguments));
-        }
-        return PLUGIN_ARGUMENTS;
-    }
-
-    private static Argument mergeArguments(final Argument a, final Argument b) {
-        if (a instanceof EnumArgument && b instanceof EnumArgument) {
-            final Set<String> vals = ImmutableSet.<String>builder()
-                    .addAll(((EnumArgument) a).values)
-                    .addAll(((EnumArgument) b).values)
-                    .build();
-            return new EnumArgument(a.property, vals);
-        } else {
-            return a;
-        }
-    }
-
-    private static Argument getArgument(Element param) {
-        final String name = param.getAttribute("name");
-        final String type = param.getAttribute("type");
-        switch (type) {
-            case "file":
-                return new FileArgument(name);
-            case "enum":
-                final Set<String> vals = getChildElements(param).stream()
-                        .map(XMLUtils::getText)
-                        .collect(Collectors.toSet());
-                if (vals.size() == 2) {
-                    for (Map.Entry<String, String> pair : TRUTHY_VALUES.entrySet()) {
-                        if (vals.contains(pair.getKey()) && vals.contains(pair.getValue())) {
-                            return new BooleanArgument(name, pair.getKey(), pair.getValue());
-                        }
-                    }
-                }
-                return new EnumArgument(name, vals);
-            default:
-                return new StringArgument(name);
-        }
-    }
-
-    private static final Map<String, String> RESERVED_PROPERTIES = ImmutableMap.of(
-            "transtype", "-f",
-            "args.input", "-i",
-            "output.dir", "-o",
-            "args.filter", "--filter",
-            ANT_TEMP_DIR, "-t"
-    );
-
-    private static final Map<String, String> TRUTHY_VALUES;
-
-    static {
-        TRUTHY_VALUES = ImmutableMap.<String, String>builder()
-                .put("true", "false")
-                .put("TRUE", "FALSE")
-                .put("yes", "no")
-                .put("YES", "NO")
-                .put("1", "0")
-                .put("on", "off")
-                .put("ON", "OFF")
-                .build();
-    }
-
-    /**
-     * The default build file name. {@value}
-     */
-    public static final String DEFAULT_BUILD_FILENAME = "build.xml";
-
-    /**
-     * Our current message output status. Follows Project.MSG_XXX.
-     */
-    private int msgOutputLevel = Project.MSG_WARN;
-
     /**
      * File that we are using for configuration.
      */
-    private File buildFile; /* null */
-    /**
-     * Run integrator
-     */
-    private boolean install;
-    /**
-     * Plug-in installation file. May be either a system path or a URL.
-     */
-    private String installFile;
-    /**
-     * Project file
-     */
-    private File projectFile;
-    /**
-     * Plug-in uninstall ID.
-     */
-    private String uninstallId;
+    private File buildFile;
 
     /**
      * Stream to use for logging.
@@ -345,70 +85,18 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
     private List<Map<String, Object>> projectProps;
 
     /**
-     * Names of classes to add as listeners to project.
-     */
-    private final Vector<String> listeners = new Vector<>(1);
-
-    /**
-     * File names of property files to load on startup.
-     */
-    private final Vector<String> propertyFiles = new Vector<>(1);
-
-    /**
-     * Indicates whether this build is to support interactive input
-     */
-    private boolean allowInput = true;
-
-    /**
-     * keep going mode
-     */
-    private boolean keepGoingMode = false;
-
-    /**
-     * The Ant logger class. There may be only one logger. It will have the
-     * right to use the 'out' PrintStream. The class must implements the
-     * BuildLogger interface.
-     */
-    private String loggerClassname = null;
-
-    /**
-     * The Ant InputHandler class. There may be only one input handler.
-     */
-    private String inputHandlerClassname = null;
-
-    /**
-     * Whether or not output to the log is to be unadorned.
-     */
-    private boolean emacsMode = false;
-
-    /**
      * Whether or not this instance has successfully been constructed and is
      * ready to run.
      */
     private boolean readyToRun = false;
 
-    /**
-     * Whether or not a logfile is being used. This is used to check if the
-     * output streams must be closed.
-     */
-    private static boolean isLogFileUsed = false;
+    private final ArgumentParser argumentParser = new ArgumentParser();
+    private Arguments args;
 
     /**
-     * optional thread priority
-     */
-    private Integer threadPriority = null;
-
-    /**
-     * proxy flag: default is false
-     */
-    private boolean proxy = false;
-
-    /**
-     * Prints the message of the Throwable if it (the message) is not
-     * <code>null</code>.
+     * Prints the message of the Throwable if it (the message) is not {@code null}.
      *
-     * @param t Throwable to print the message of. Must not be <code>null</code>
-     *          .
+     * @param t Throwable to print the message of. Must not be {@code null}
      */
     private void printMessage(final Throwable t) {
         final String message = t.getMessage();
@@ -418,7 +106,7 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
     }
 
     private void printErrorMessage(final String msg) {
-        if (useColor) {
+        if (args.useColor) {
             System.err.print(DefaultLogger.ANSI_RED);
             System.err.print("Error: " + msg);
             System.err.println(DefaultLogger.ANSI_RESET);
@@ -490,7 +178,6 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
             }
         }
 
-
         // expect the worst
         int exitCode = 1;
         try {
@@ -534,8 +221,8 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
      *
      * @since Ant 1.6
      */
-    private static void handleLogfile() {
-        if (isLogFileUsed) {
+    private void handleLogfile() {
+        if (args.logFile != null) {
             FileUtils.close(out);
             FileUtils.close(err);
         }
@@ -566,123 +253,28 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
      * @since Ant 1.6
      */
     private void processArgs(final String[] arguments) {
-        PrintStream logTo = null;
+        args = argumentParser.processArgs(arguments);
+        final Map<String, Object> definedProps = new HashMap<>(args.definedProps);
+        buildFile = args.buildFile;
 
-        // cycle through given args
-
-        boolean justPrintUsage = false;
-        boolean justPrintVersion = false;
-        boolean justPrintDiagnostics = false;
-        boolean justPrintPlugins = false;
-        boolean justPrintTranstypes = false;
-        boolean justPrintDeliverables = false;
-        useColor = getUseColor();
-
-        final Map<String, Object> definedProps = new HashMap<>();
-        final Deque<String> args = new ArrayDeque<>(Arrays.asList(arguments));
-        while (!args.isEmpty()) {
-            final String arg = args.pop();
-
-            if (isLongForm(arg, "-help") || arg.equals("-h")) {
-                justPrintUsage = true;
-            } else if (isLongForm(arg, "-version")) {
-                justPrintVersion = true;
-            } else if (isLongForm(arg, "-plugins")) {
-                justPrintPlugins = true;
-            } else if (isLongForm(arg, "-transtypes")) {
-                justPrintTranstypes = true;
-            } else if (isLongForm(arg, "-deliverables")) {
-                justPrintDeliverables = true;
-            } else if (isLongForm(arg, "-install")) {
-                handleArgInstall(arg, args);
-            } else if (isLongForm(arg, "-project") || arg.equals("-p")) {
-                handleArgProject(arg, args);
-            } else if (isLongForm(arg, "-force")) {
-                definedProps.put("force", "true");
-            } else if (isLongForm(arg, "-uninstall")) {
-                handleArgUninstall(arg, args);
-            } else if (isLongForm(arg, "-diagnostics")) {
-                justPrintDiagnostics = true;
-                // } else if (arg.equals("-quiet") || arg.equals("-q")) {
-                // msgOutputLevel = Project.MSG_WARN;
-            } else if (isLongForm(arg, "-verbose") || arg.equals("-v")) {
-                msgOutputLevel = Project.MSG_INFO;
-            } else if (isLongForm(arg, "-debug") || arg.equals("-d")) {
-                msgOutputLevel = Project.MSG_VERBOSE;
-            } else if (isLongForm(arg, "-noinput")) {
-                allowInput = false;
-            } else if (isLongForm(arg, "-logfile") || arg.equals("-l")) {
-                logTo = handleArgLogFile(args);
-            } else if (isLongForm(arg, "-buildfile") || isLongForm(arg, "-file")) { //|| arg.equals("-f")
-                handleArgBuildFile(args);
-            } else if (isLongForm(arg, "-listener")) {
-                handleArgListener(args);
-            } else if (arg.startsWith("-D")) {
-                handleArgDefine(arg, args);
-            } else if (isLongForm(arg, "-logger")) {
-                handleArgLogger(args);
-            } else if (isLongForm(arg, "-inputhandler")) {
-                handleArgInputHandler(args);
-            } else if (isLongForm(arg, "-emacs") || arg.equals("-e")) {
-                emacsMode = true;
-            } else if (isLongForm(arg, "-propertyfile")) {
-                handleArgPropertyFile(arg, args);
-            } else if (arg.equals("-k") || isLongForm(arg, "-keep-going")) {
-                keepGoingMode = true;
-            } else if (isLongForm(arg, "-nice")) {
-                handleArgNice(args);
-            } else if (ARGUMENTS.containsKey(getArgumentName(arg))) {
-                definedProps.putAll(handleParameterArg(arg, args, ARGUMENTS.get(getArgumentName(arg))));
-            } else if (getPluginArguments().containsKey(getArgumentName(arg))) {
-                definedProps.putAll(handleParameterArg(arg, args, getPluginArguments().get(getArgumentName(arg))));
-            } else if (LAUNCH_COMMANDS.contains(arg)) {
-                // catch script/ant mismatch with a meaningful message
-                // we could ignore it, but there are likely to be other
-                // version problems, so we stamp down on the configuration now
-                final String msg = "Ant's Main method is being handed " + "an option " + arg
-                        + " that is only for the launcher class."
-                        + "\nThis can be caused by a version mismatch between "
-                        + "the ant script/.bat file and Ant itself.";
-                throw new BuildException(msg);
-            } else if (isLongForm(arg, "-autoproxy")) {
-                proxy = true;
-            } else if (arg.startsWith("-") || arg.startsWith("/")) {
-                // we don't have any more args to recognize!
-                final String msg = "Error: Unknown argument: " + arg;
-                printErrorMessage(msg);
-                printUsage();
-                throw new BuildException("");
-            } else {
-                // if it's no other arg, it may be the target
-                targets.addElement(arg);
+        if (args.justPrintUsage || args.justPrintVersion || args.justPrintDiagnostics || args.justPrintPlugins || args.justPrintTranstypes || args.justPrintDeliverables) {
+            if (args.justPrintVersion) {
+                printVersion(args.msgOutputLevel);
             }
-        }
-
-        if (install && msgOutputLevel < Project.MSG_INFO) {
-            emacsMode = true;
-        }
-
-        // Load the property files specified by --propertyfile
-        definedProps.putAll(loadPropertyFiles());
-
-        if (justPrintUsage || justPrintVersion || justPrintDiagnostics || justPrintPlugins || justPrintTranstypes || justPrintDeliverables) {
-            if (justPrintVersion) {
-                printVersion(msgOutputLevel);
-            }
-            if (justPrintUsage) {
+            if (args.justPrintUsage) {
                 printUsage();
             }
-            if (justPrintDiagnostics) {
-                Diagnostics.doReport(System.out, msgOutputLevel);
+            if (args.justPrintDiagnostics) {
+                Diagnostics.doReport(System.out, args.msgOutputLevel);
             }
-            if (justPrintPlugins) {
+            if (args.justPrintPlugins) {
                 printPlugins();
             }
-            if (justPrintTranstypes) {
+            if (args.justPrintTranstypes) {
                 printTranstypes();
             }
-            if (justPrintDeliverables) {
-                if (projectFile == null) {
+            if (args.justPrintDeliverables) {
+                if (args.projectFile == null) {
                     printErrorMessage("Error: Project file not defined");
                     printUsage();
                     throw new BuildException("");
@@ -691,25 +283,27 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
             }
 
             return;
-        } else if (install) {
+        }
+
+        if (args.install) {
             buildFile = findBuildFile(System.getProperty("dita.dir"), "integrator.xml");
             targets.clear();
-            if (installFile != null) {
+            if (args.installFile != null) {
                 targets.add("install");
-                final File f = new File(installFile.replace('/', File.separatorChar)).getAbsoluteFile();
+                final File f = new File(args.installFile.replace('/', File.separatorChar)).getAbsoluteFile();
                 if (f.exists()) {
                     definedProps.put("plugin.file", f.getAbsolutePath());
                 } else {
-                    definedProps.put("plugin.file", installFile);
+                    definedProps.put("plugin.file", args.installFile);
                 }
-            } else if (uninstallId != null) {
+            } else if (args.uninstallId != null) {
                 targets.add("uninstall");
-                definedProps.put("plugin.id", uninstallId);
+                definedProps.put("plugin.id", args.uninstallId);
             } else {
                 targets.add("integrate");
             }
         } else {
-            if (projectFile == null) {
+            if (args.projectFile == null) {
                 if (!definedProps.containsKey("transtype")) {
                     printErrorMessage("Error: Transformation type not defined");
                     printUsage();
@@ -732,12 +326,12 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
             }
         }
 
-        projectProps = projectFile != null
+        projectProps = args.projectFile != null
                 ? handleProject(definedProps)
                 : Collections.singletonList(definedProps);
 
         // make sure buildfile exists
-        if (!buildFile.exists() || buildFile.isDirectory()) {
+        if (!args.buildFile.exists() || buildFile.isDirectory()) {
             System.out.println("Buildfile: " + buildFile + " does not exist!");
             throw new BuildException("Build failed");
         }
@@ -745,11 +339,18 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
         // Normalize buildFile for re-import detection
         buildFile = FileUtils.getFileUtils().normalize(buildFile.getAbsolutePath());
 
-        if (msgOutputLevel >= Project.MSG_VERBOSE) {
+        if (args.msgOutputLevel >= Project.MSG_VERBOSE) {
             System.out.println("Buildfile: " + buildFile);
         }
 
-        if (logTo != null) {
+        if (args.logFile != null) {
+            PrintStream logTo;
+            try {
+                logTo = new PrintStream(new FileOutputStream(args.logFile));
+            } catch (final IOException ioe) {
+                throw new BuildException("Cannot write on the specified log file. "
+                        + "Make sure the path exists and you have write permissions.");
+            }
             out = logTo;
             err = logTo;
             System.setOut(out);
@@ -759,7 +360,7 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
     }
 
     private List<Map<String, Object>> handleProject(Map<String, Object> definedProps) {
-        final URI base = projectFile.toURI();
+        final URI base = args.projectFile.toURI();
         final org.dita.dost.project.Project project = readProjectFile();
         final String runDeliverable = (String) definedProps.get("project.deliverable");
 
@@ -783,12 +384,12 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
                             return;
                         }
                         if (param.value != null) {
-                            final Argument argument = getPluginArguments().getOrDefault(param.name, new StringArgument(param.name));
+                            final Argument argument = argumentParser.getPluginArguments().getOrDefault(param.name, new StringArgument(param.name));
                             final String value = argument.getValue(param.value);
                             props.put(param.name, value);
                         } else {
                             final String value;
-                            final Argument argument = getPluginArguments().get("--" + param.name);
+                            final Argument argument = argumentParser.getPluginArguments().get("--" + param.name);
                             if (argument != null && (argument instanceof FileArgument || argument instanceof AbsoluteFileArgument)) {
                                 value = Paths.get(base.resolve(param.href)).toString();
                             } else {
@@ -816,118 +417,16 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
     }
 
     private org.dita.dost.project.Project readProjectFile() throws BuildException {
-        if (!projectFile.exists()) {
-            printErrorMessage("Project file " + projectFile + " does not exist");
+        if (!args.projectFile.exists()) {
+            printErrorMessage("Project file " + args.projectFile + " does not exist");
             throw new BuildException("");
         }
         try {
-            return ProjectFactory.load(projectFile.toURI());
+            return ProjectFactory.load(args.projectFile.toURI());
         } catch (Exception e) {
             printErrorMessage(e.getMessage());
             throw new BuildException("");
         }
-    }
-
-    private boolean getUseColor() {
-        final String os = System.getProperty("os.name");
-        if (os != null && os.startsWith("Windows")) {
-            return false;
-        }
-        return Boolean.parseBoolean(Configuration.configuration.getOrDefault("cli.color", "true"));
-    }
-
-    private PrintStream handleArgLogFile(Deque<String> args) {
-        PrintStream logTo;
-        final String value = args.pop();
-        try {
-            final File logFile = new File(value);
-            logTo = new PrintStream(new FileOutputStream(logFile));
-            isLogFileUsed = true;
-            useColor = false;
-        } catch (final IOException ioe) {
-            final String msg = "Cannot write on the specified log file. "
-                    + "Make sure the path exists and you have write " + "permissions.";
-            throw new BuildException(msg);
-        } catch (final ArrayIndexOutOfBoundsException aioobe) {
-            final String msg = "You must specify a log file when " + "using the --log argument";
-            throw new BuildException(msg);
-        }
-        return logTo;
-    }
-
-    private boolean isLongForm(String arg, String property) {
-        final String name = arg.contains("=") ? arg.substring(0, arg.indexOf('=')) : arg;
-        return name.equals(property) || name.equals("-" + property);
-    }
-
-    // --------------------------------------------------------
-    // Methods for handling the command line arguments
-    // --------------------------------------------------------
-
-    /**
-     * Handle the --install argument
-     */
-    private void handleArgInstall(final String arg, final Deque<String> args) {
-        install = true;
-        String name = arg;
-        final int posEq = name.indexOf("=");
-        String value;
-        if (posEq != -1) {
-            value = name.substring(posEq + 1);
-        } else {
-            value = args.peek();
-            if (value != null && !value.startsWith("-")) {
-                value = args.pop();
-            } else {
-                value = null;
-            }
-        }
-        if (value != null) {
-            installFile = value;
-        }
-    }
-
-    /**
-     * Handle the --project argument
-     */
-    private void handleArgProject(final String arg, final Deque<String> args) {
-        String name = arg;
-        final int posEq = name.indexOf("=");
-        final String value;
-        if (posEq != -1) {
-            value = name.substring(posEq + 1);
-            name = name.substring(0, posEq);
-        } else {
-            value = args.pop();
-        }
-        if (value == null) {
-            throw new BuildException("Missing value for project " + name);
-        }
-        projectFile = new File(value).getAbsoluteFile();
-    }
-
-    /**
-     * Handle the --uninstall argument
-     */
-    private void handleArgUninstall(final String arg, final Deque<String> args) {
-        install = true;
-        String name = arg;
-        final int posEq = name.indexOf("=");
-        String value;
-        if (posEq != -1) {
-            value = name.substring(posEq + 1);
-        } else {
-            value = args.peek();
-            if (value != null && !value.startsWith("-")) {
-                value = args.pop();
-            } else {
-                value = null;
-            }
-        }
-        if (value == null) {
-            throw new BuildException("You must specify a installation package when using the --uninstall argument");
-        }
-        uninstallId = value;
     }
 
     /**
@@ -969,189 +468,6 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
     }
 
     /**
-     * Handle the --buildfile, --file, -f argument
-     */
-    private void handleArgBuildFile(final Deque<String> args) {
-        final String value = args.pop();
-        if (value == null) {
-            throw new BuildException("You must specify a buildfile when using the --buildfile argument");
-        }
-        buildFile = new File(value.replace('/', File.separatorChar));
-    }
-
-    /**
-     * Handle --listener argument
-     */
-    private void handleArgListener(final Deque<String> args) {
-        final String value = args.pop();
-        if (value == null) {
-            throw new BuildException("You must specify a classname when using the --listener argument");
-        }
-        listeners.addElement(value);
-    }
-
-    /**
-     * Handler -D argument
-     */
-    private Map<String, Object> handleArgDefine(final String arg, final Deque<String> args) {
-        /*
-         * Interestingly enough, we get to here when a user uses -Dname=value.
-         * However, in some cases, the OS goes ahead and parses this out to args
-         * {"-Dname", "value"} so instead of parsing on "=", we just make the
-         * "-D" characters go away and skip one argument forward.
-         *
-         * I don't know how to predict when the JDK is going to help or not, so
-         * we simply look for the equals sign.
-         */
-        String name = arg.substring(2);
-        String value;
-        final int posEq = name.indexOf("=");
-        if (posEq > 0) {
-            value = name.substring(posEq + 1);
-            name = name.substring(0, posEq);
-        } else {
-            value = args.pop();
-        }
-        if (value == null) {
-            throw new BuildException("Missing value for property " + name);
-        }
-
-        if (RESERVED_PROPERTIES.containsKey(name)) {
-            throw new BuildException("Property " + name + " cannot be set with -D, use " + RESERVED_PROPERTIES.get(name) + " instead");
-        }
-        return ImmutableMap.of(name, value);
-    }
-
-    /**
-     * Handler parameter argument
-     */
-    private Map<String, Object> handleParameterArg(final String arg, final Deque<String> args, final Argument argument) {
-        String name = arg;
-        String value;
-        final int posEq = name.indexOf("=");
-        if (posEq > 0) {
-            value = name.substring(posEq + 1);
-            name = name.substring(0, posEq);
-        } else {
-            value = args.pop();
-        }
-        if (value == null) {
-            throw new BuildException("Missing value for property " + name);
-        }
-        return ImmutableMap.of(argument.property, argument.getValue(value));
-    }
-
-    /**
-     * Get argument name
-     */
-    private String getArgumentName(final String arg) {
-        int pos = arg.indexOf("=");
-        if (pos == -1) {
-            pos = arg.indexOf(":");
-        }
-        return arg.substring(0, pos != -1 ? pos : arg.length());
-    }
-
-    /**
-     * Handle the --logger argument.
-     */
-    private void handleArgLogger(final Deque<String> args) {
-        if (loggerClassname != null) {
-            throw new BuildException("Only one logger class may be specified.");
-        }
-        loggerClassname = args.pop();
-        if (loggerClassname == null) {
-            throw new BuildException("You must specify a classname when using the -logger argument");
-        }
-    }
-
-    /**
-     * Handle the --inputhandler argument.
-     */
-    private void handleArgInputHandler(final Deque<String> args) {
-        if (inputHandlerClassname != null) {
-            throw new BuildException("Only one input handler class may be specified.");
-        }
-        inputHandlerClassname = args.pop();
-        if (inputHandlerClassname == null) {
-            throw new BuildException("You must specify a classname when using the --inputhandler" + " argument");
-        }
-    }
-
-    /**
-     * Handle the --propertyfile argument.
-     */
-    private void handleArgPropertyFile(final String arg, final Deque<String> args) {
-        String name = arg.substring(2);
-        String value;
-        final int posEq = name.indexOf("=");
-        if (posEq > 0) {
-            value = name.substring(posEq + 1);
-        } else {
-            value = args.pop();
-        }
-        if (value == null) {
-            throw new BuildException("You must specify a property filename when using the --propertyfile argument");
-        }
-        propertyFiles.addElement(value);
-    }
-
-    /**
-     * Handle the --nice argument.
-     */
-    private void handleArgNice(final Deque<String> args) {
-        final String value = args.pop();
-        if (value == null) {
-            throw new BuildException("You must supply a niceness value (1-10) after the --nice option");
-        }
-        try {
-            threadPriority = Integer.decode(value);
-        } catch (final NumberFormatException e) {
-            throw new BuildException("Unrecognized niceness value: " + value);
-        }
-
-        if (threadPriority < Thread.MIN_PRIORITY || threadPriority > Thread.MAX_PRIORITY) {
-            throw new BuildException("Niceness value is out of the range 1-10");
-        }
-    }
-
-    /**
-     * Load the property files specified by --propertyfile
-     */
-    private Map<String, Object> loadPropertyFiles() {
-        final Map<String, Object> definedProps = new HashMap<>();
-        for (int propertyFileIndex = 0; propertyFileIndex < propertyFiles.size(); propertyFileIndex++) {
-            final String filename = propertyFiles.elementAt(propertyFileIndex);
-            final Properties props = new Properties();
-            FileInputStream fis = null;
-            try {
-                fis = new FileInputStream(filename);
-                props.load(fis);
-            } catch (final IOException e) {
-                System.out.println("Could not load property file " + filename + ": " + e.getMessage());
-            } finally {
-                FileUtils.close(fis);
-            }
-
-            // ensure that -D properties take precedence
-            final Enumeration propertyNames = props.propertyNames();
-            while (propertyNames.hasMoreElements()) {
-                final String name = propertyNames.nextElement().toString();
-                if (!definedProps.containsKey(name)) {
-                    final Argument arg = getPluginArguments().get("--" + name);
-                    final String value = props.getProperty(name);
-                    if (arg != null) {
-                        definedProps.put(name, arg.getValue(value));
-                    } else {
-                        definedProps.put(name, value);
-                    }
-                }
-            }
-        }
-        return definedProps;
-    }
-
-    /**
      * Helper to get the parent file for a given file.
      * <p>
      * Added to simulate File.getParentFile() from JDK 1.2.
@@ -1164,7 +480,7 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
     private File getParentFile(final File file) {
         final File parent = file.getParentFile();
 
-        if (parent != null && msgOutputLevel >= Project.MSG_VERBOSE) {
+        if (parent != null && args.msgOutputLevel >= Project.MSG_VERBOSE) {
             System.out.println("Searching in " + parent.getAbsolutePath());
         }
 
@@ -1185,7 +501,7 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
      * not
      */
     private File findBuildFile(final String start, final String suffix) {
-        if (msgOutputLevel >= Project.MSG_INFO) {
+        if (args.msgOutputLevel >= Project.MSG_INFO) {
             System.out.println("Searching for " + suffix + " ...");
         }
 
@@ -1242,7 +558,7 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
             // ant class if we are going to implement it.
             // System.setSecurityManager(new NoExitSecurityManager());
             try {
-                if (allowInput) {
+                if (args.allowInput) {
                     project.setDefaultInputStream(System.in);
                 }
                 System.setIn(new DemuxInputStream(project));
@@ -1252,10 +568,10 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
                 project.fireBuildStarted();
 
                 // set the thread priorities
-                if (threadPriority != null) {
+                if (args.threadPriority != null) {
                     try {
-                        project.log("Setting Ant's thread priority to " + threadPriority, Project.MSG_VERBOSE);
-                        Thread.currentThread().setPriority(threadPriority);
+                        project.log("Setting Ant's thread priority to " + args.threadPriority, Project.MSG_VERBOSE);
+                        Thread.currentThread().setPriority(args.threadPriority);
                     } catch (final SecurityException swallowed) {
                         // we cannot set the priority here.
                         project.log("A security manager refused to set the -nice value");
@@ -1280,8 +596,8 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
                 project.setUserProperty(MagicNames.ANT_FILE, buildFile.getAbsolutePath());
                 project.setUserProperty(MagicNames.ANT_FILE_TYPE, MagicNames.ANT_FILE_TYPE_FILE);
 
-                project.setKeepGoingMode(keepGoingMode);
-                if (proxy) {
+                project.setKeepGoingMode(args.keepGoingMode);
+                if (args.proxy) {
                     // proxy setup if enabled
                     final ProxySetup proxySetup = new ProxySetup(project);
                     proxySetup.enableProxies();
@@ -1337,13 +653,12 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
      */
     @Override
     protected void addBuildListeners(final Project project) {
-
         // Add the default listener
         project.addBuildListener(createLogger());
 
-        final int count = listeners.size();
+        final int count = args.listeners.size();
         for (int i = 0; i < count; i++) {
-            final String className = listeners.elementAt(i);
+            final String className = args.listeners.elementAt(i);
             final BuildListener listener = ClasspathUtils.newInstance(className,
                     Main.class.getClassLoader(), BuildListener.class);
             project.setProjectReference(listener);
@@ -1360,11 +675,11 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
      *                        could not be loaded.
      */
     private void addInputHandler(final Project project) throws BuildException {
-        InputHandler handler;
-        if (inputHandlerClassname == null) {
+        final InputHandler handler;
+        if (args.inputHandlerClassname == null) {
             handler = new DefaultInputHandler();
         } else {
-            handler = ClasspathUtils.newInstance(inputHandlerClassname, Main.class.getClassLoader(),
+            handler = ClasspathUtils.newInstance(args.inputHandlerClassname, Main.class.getClassLoader(),
                     InputHandler.class);
             project.setProjectReference(handler);
         }
@@ -1383,24 +698,24 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
      */
     private BuildLogger createLogger() {
         BuildLogger logger;
-        if (loggerClassname != null) {
+        if (args.loggerClassname != null) {
             try {
-                logger = ClasspathUtils.newInstance(loggerClassname, Main.class.getClassLoader(),
+                logger = ClasspathUtils.newInstance(args.loggerClassname, Main.class.getClassLoader(),
                         BuildLogger.class);
             } catch (final BuildException e) {
-                printErrorMessage("The specified logger class " + loggerClassname + " could not be used because "
+                printErrorMessage("The specified logger class " + args.loggerClassname + " could not be used because "
                         + e.getMessage());
                 throw new RuntimeException();
             }
         } else {
             logger = new DefaultLogger();
-            ((DefaultLogger) logger).useColor(useColor);
+            ((DefaultLogger) logger).useColor(args.useColor);
         }
 
-        logger.setMessageOutputLevel(msgOutputLevel);
+        logger.setMessageOutputLevel(args.msgOutputLevel);
         logger.setOutputPrintStream(out);
         logger.setErrorPrintStream(err);
-        logger.setEmacsMode(emacsMode);
+        logger.setEmacsMode(args.emacsMode);
 
         return logger;
     }
@@ -1491,191 +806,5 @@ public class Main extends org.apache.tools.ant.Main implements AntMain {
     private static void printVersion(final int logLevel) throws BuildException {
         System.out.println("DITA-OT version " + Configuration.configuration.get("otversion"));
         // System.out.println(getAntVersion());
-    }
-
-    /**
-     * Prints the description of a project (if there is one) to
-     * <code>System.out</code>.
-     *
-     * @param project The project to display a description of. Must not be
-     *                <code>null</code>.
-     */
-    private static void printDescription(final Project project) {
-        if (project.getDescription() != null) {
-            project.log(project.getDescription());
-        }
-    }
-
-    /**
-     * Targets in imported files with a project name and not overloaded by the
-     * main build file will be in the target map twice. This method removes the
-     * duplicate target.
-     *
-     * @param targets the targets to filter.
-     * @return the filtered targets.
-     */
-    private static Map<String, Target> removeDuplicateTargets(final Map<String, Target> targets) {
-        final Map<Location, Target> locationMap = new HashMap<>();
-        for (final Map.Entry<String, Target> entry : targets.entrySet()) {
-            final String name = entry.getKey();
-            final Target target = entry.getValue();
-            final Target otherTarget = locationMap.get(target.getLocation());
-            // Place this entry in the location map if
-            // a) location is not in the map
-            // b) location is in map, but it's name is longer
-            // (an imported target will have a name. prefix)
-            if (otherTarget == null || otherTarget.getName().length() > name.length()) {
-                locationMap.put(target.getLocation(), target); // Smallest name
-                // wins
-            }
-        }
-        final Map<String, Target> ret = new HashMap<>();
-        for (final Target target : locationMap.values()) {
-            ret.put(target.getName(), target);
-        }
-        return ret;
-    }
-
-    /**
-     * Prints a list of all targets in the specified project to
-     * <code>System.out</code>, optionally including subtargets.
-     *
-     * @param project         The project to display a description of. Must not be
-     *                        <code>null</code>.
-     * @param printSubTargets Whether or not subtarget names should also be
-     *                        printed.
-     */
-    private static void printTargets(final Project project, boolean printSubTargets, final boolean printDependencies) {
-        // find the target with the longest name
-        int maxLength = 0;
-        final Map<String, Target> ptargets = removeDuplicateTargets(project.getTargets());
-        String targetName;
-        String targetDescription;
-        Target currentTarget;
-        // split the targets in top-level and sub-targets depending
-        // on the presence of a description
-        final Vector<String> topNames = new Vector<>();
-        final Vector<String> topDescriptions = new Vector<>();
-        final Vector<Enumeration<String>> topDependencies = new Vector<>();
-        final Vector<String> subNames = new Vector<>();
-        final Vector<Enumeration<String>> subDependencies = new Vector<>();
-
-        for (Target target : ptargets.values()) {
-            currentTarget = target;
-            targetName = currentTarget.getName();
-            if (targetName.equals("")) {
-                continue;
-            }
-            targetDescription = currentTarget.getDescription();
-            // maintain a sorted list of targets
-            if (targetDescription == null) {
-                final int pos = findTargetPosition(subNames, targetName);
-                subNames.insertElementAt(targetName, pos);
-                if (printDependencies) {
-                    subDependencies.insertElementAt(currentTarget.getDependencies(), pos);
-                }
-            } else {
-                final int pos = findTargetPosition(topNames, targetName);
-                topNames.insertElementAt(targetName, pos);
-                topDescriptions.insertElementAt(targetDescription, pos);
-                if (targetName.length() > maxLength) {
-                    maxLength = targetName.length();
-                }
-                if (printDependencies) {
-                    topDependencies.insertElementAt(currentTarget.getDependencies(), pos);
-                }
-            }
-        }
-
-        printTargets(project, topNames, topDescriptions, topDependencies, "Main targets:", maxLength);
-        // if there were no main targets, we list all subtargets
-        // as it means nothing has a description
-        if (topNames.size() == 0) {
-            printSubTargets = true;
-        }
-        if (printSubTargets) {
-            printTargets(project, subNames, null, subDependencies, "Other targets:", 0);
-        }
-
-        final String defaultTarget = project.getDefaultTarget();
-        if (defaultTarget != null && !"".equals(defaultTarget)) {
-            // shouldn't need to check but...
-            project.log("Default target: " + defaultTarget);
-        }
-    }
-
-    /**
-     * Searches for the correct place to insert a name into a list so as to keep
-     * the list sorted alphabetically.
-     *
-     * @param names The current list of names. Must not be <code>null</code>.
-     * @param name  The name to find a place for. Must not be <code>null</code>.
-     * @return the correct place in the list for the given name
-     */
-    private static int findTargetPosition(final Vector<String> names, final String name) {
-        final int size = names.size();
-        int res = size;
-        for (int i = 0; i < size && res == size; i++) {
-            if (name.compareTo(names.elementAt(i)) < 0) {
-                res = i;
-            }
-        }
-        return res;
-    }
-
-    /**
-     * Writes a formatted list of target names to <code>System.out</code> with
-     * an optional description.
-     *
-     * @param project      the project instance.
-     * @param names        The names to be printed. Must not be <code>null</code>.
-     * @param descriptions The associated target descriptions. May be
-     *                     <code>null</code>, in which case no descriptions are
-     *                     displayed. If non-<code>null</code>, this should have as many
-     *                     elements as <code>names</code>.
-     * @param dependencies The list of dependencies for each target. The
-     *                     dependencies are listed as a non null enumeration of String.
-     * @param heading      The heading to display. Should not be <code>null</code>.
-     * @param maxlen       The maximum length of the names of the targets. If
-     *                     descriptions are given, they are padded to this position so
-     *                     they line up (so long as the names really <i>are</i> shorter
-     *                     than this).
-     */
-    private static void printTargets(final Project project, final Vector<String> names,
-                                     final Vector<String> descriptions, final Vector<Enumeration<String>> dependencies, final String heading,
-                                     final int maxlen) {
-        // now, start printing the targets and their descriptions
-        final String lSep = System.getProperty("line.separator");
-        // got a bit annoyed that I couldn't find a pad function
-        String spaces = "    ";
-        while (spaces.length() <= maxlen) {
-            spaces += spaces;
-        }
-        final StringBuilder msg = new StringBuilder();
-        msg.append(heading).append(lSep).append(lSep);
-        final int size = names.size();
-        for (int i = 0; i < size; i++) {
-            msg.append(" ");
-            msg.append(names.elementAt(i));
-            if (descriptions != null) {
-                msg.append(spaces, 0, maxlen - names.elementAt(i).length() + 2);
-                msg.append(descriptions.elementAt(i));
-            }
-            msg.append(lSep);
-            if (!dependencies.isEmpty()) {
-                final Enumeration<String> deps = dependencies.elementAt(i);
-                if (deps.hasMoreElements()) {
-                    msg.append("   depends on: ");
-                    while (deps.hasMoreElements()) {
-                        msg.append(deps.nextElement());
-                        if (deps.hasMoreElements()) {
-                            msg.append(", ");
-                        }
-                    }
-                    msg.append(lSep);
-                }
-            }
-        }
-        project.log(msg.toString(), Project.MSG_WARN);
     }
 }
